# swagger.yaml
openapi: 3.0.0
info:
  title: 'API communication documentation'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:9000'
    description: 'Local development server'
paths:
  /v1/auth/login:
    post:
      tags:
        - auth
      summary: login
      description: 'This service is for login.'
      operationId: 9e741540f6f7dad0dfdbb05f7c817532
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  title: 'User Email'
                  type: string
                  example: example@gmail.com
                password:
                  title: 'User password'
                  type: string
                  example: '@A48que95#'
              type: object
      responses:
        '200':
          description: 'The operation was successful.'
          content:
            application/json:
              schema:
                properties:
                  token: { title: 'jwt token', type: string, example: eyJ0eXAiOiJKV1QiLqwCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJkb2qwetvLmlyIiwiYXVkIjoic2VsbGVyIiwiaWF0IjoxNjQ4MDIyOTE3LCJleHAiOjE3MzQ0MjI5MTcsInN1YiI6MX0.r_UgDmle7TN2xlnlXVmmeyfoH_MW_4Awc3YDOcoYjgk }
                  expired_at: { title: 'Token expiration date', type: string, example: '2025-11-28 16:14:45' }
                type: object
        '404':
          description: 'The password is incorrect or the username does not exist.'
        '422':
          description: 'The information submitted is invalid.'
  /v1/auth/register:
    post:
      tags:
        - auth
      summary: registration
      description: 'This service is for registration.'
      operationId: 2a8d94cdfa0e4c500eedc9e32a2aa1b9
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - first_name
                - email
                - password
              properties:
                first_name:
                  title: 'User first name'
                  type: string
                  example: Bob
                last_name:
                  title: 'User last name'
                  type: string
                  example: Alice
                email:
                  title: 'User Email'
                  type: string
                  example: example@gmail.com
                password:
                  title: 'User password'
                  type: string
                  example: '@A48que95#'
              type: object
      responses:
        '201':
          description: 'Registration was successful.'
        '400':
          description: 'You are already registered.'
        '422':
          description: 'The information submitted is invalid.'
  '/v1/evaluations/{evaluation}':
    delete:
      tags:
        - evaluations
      summary: delete
      description: 'This service is for evaluation delete.'
      operationId: dde8491cfc07274fd5609dc09bf5e66d
      parameters:
        -
          name: evaluation
          in: path
          description: 'evaluation id'
          required: true
          schema:
            type: string
            example: '1'
      responses:
        '200':
          description: 'The operation was successful.'
        '404':
          description: 'Lack of access or non-existence of assessment.'
        '422':
          description: 'The information submitted is invalid.'
      security:
        -
          bearerAuth: []
    patch:
      tags:
        - evaluations
      summary: update
      description: 'This service is for evaluation update.'
      operationId: 7ee69c219e23954c955662a5dc5dcde6
      parameters:
        -
          name: evaluation
          in: path
          description: 'evaluation id'
          required: true
          schema:
            type: string
            example: '1'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - title
                - description
              properties:
                title:
                  title: 'evaluation title'
                  type: string
                  example: 'evaluation title'
                description:
                  title: 'evaluation description'
                  type: string
                  example: 'evaluation description'
              type: object
      responses:
        '200':
          description: 'The operation was successful.'
        '404':
          description: 'Lack of access or non-existence of assessment.'
        '422':
          description: 'The information submitted is invalid.'
      security:
        -
          bearerAuth: []
  /v1/evaluations:
    get:
      tags:
        - evaluations
      summary: 'evaluations list'
      description: 'This service is for getting a list of evaluations.'
      operationId: f5ec4adcbd8e1bef7999654adbe08eda
      parameters:
        -
          name: page
          in: query
          description: 'Page number'
          required: false
          schema:
            type: string
            example: '1'
      responses:
        '200':
          description: 'The operation was successful.'
          content:
            application/json:
              schema:
                anyOf:
                  - { title: evaluations, properties: { data: { title: 'evaluations list', type: array, items: { properties: { id: { title: 'evaluation id', type: integer, example: 1 }, title: { title: 'evaluation title', type: string, example: alice }, description: { title: 'evaluation description', type: string, example: bob }, user: { title: 'User registering the assessment', properties: { first_name: { title: 'first name', type: string, example: Amir }, last_name: { title: 'last name', type: string, example: Faramarzi, nullable: true }, email: { title: email, type: string, example: amir@gmail.com } }, type: object, nullable: true } }, type: object } }, meta: { title: 'Other information', properties: { isLastPage: { title: 'Is this page the last page in the list of desired data?', type: boolean, example: false }, perPage: { title: 'Number of data returned', type: integer, example: 15 } }, type: object } }, type: object }
              examples:
                'evaluations list':
                  summary: 'evaluations list'
                  description: 'Get a list of evaluations'
                  value: { data: [{ id: 1, title: 'evaluation title', description: 'evaluation description', user: { first_name: Amir, last_name: Faramarzi, email: amir@gmail.com }, created_at: '2025-11-27 17:29:54' }], meta: { isLastPage: false, perPage: 15 } }
      security:
        -
          bearerAuth: []
    post:
      tags:
        - evaluations
      summary: insert
      description: 'This service is for evaluation insert.'
      operationId: fcbd3442db2fbf2f5cb2f1ab47ff4c92
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - title
                - description
              properties:
                title:
                  title: 'evaluation title'
                  type: string
                  example: 'evaluation title'
                description:
                  title: 'evaluation description'
                  type: string
                  example: 'evaluation description'
              type: object
      responses:
        '201':
          description: 'The operation was successful.'
        '422':
          description: 'The information submitted is invalid.'
      security:
        -
          bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
tags:
  -
    name: auth
    description: auth
  -
    name: evaluations
    description: evaluations